Re-organize Sample_ConnectName; Add Sample_ConnectName/Sample_UI;
