+++ toio-sdk-unity/Assets/toio-sdk/Samples/Sample_ConnectName/Sample_BasicUI/README.md.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 1a0647788f9581c45ac7f1ac237c48b5
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
