+++ toio-sdk-unity/Assets/toio-sdk/Samples/Sample_ConnectName/Sample_Property.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 07df9087f4329c946b2e703a26f1df9e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
