+++ toio-sdk-unity/Assets/toio-sdk/Samples/Sample_ConnectName/Sample_Property/README.md
@@ -1,0 +1,3 @@
+## Sample_Property
+
+このサンプルは、事前にUnityエディタのプロパティで接続するキューブの Local Name を設定し、実行時にスキャンできた際に接続するサンプルです。
