+++ toio-sdk-unity/Assets/toio-sdk/Samples/Sample_ConnectName/Sample_UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bb120ed5c87250f4f86c4f92a388f7c5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
