+++ toio-sdk-unity/Assets/toio-sdk/Samples/Sample_ConnectName/Sample_UI/Sample_UI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 81b8ff2417175c54c848c0f8b7e46102
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
