+++ toio-sdk-unity/Assets/toio-sdk/Samples/Sample_ConnectName/Sample_UI/SimUI_Resizer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a6ac2e8d6d3504a4583e69da881f83bc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
